tx: make tests compile tx: temp fix circular dependency
