+++ packages/tx/src/types.ts
@@ -1,9 +1,7 @@
 import { AddressLike, BN, BNLike, BufferLike } from 'ethereumjs-util'
 import Common from '@ethereumjs/common'
-import { default as UnsignedLegacyTransaction } from './unsignedLegacyTransaction'
-import { default as UnsignedEIP2930Transaction } from './unsignedEIP2930Transaction'
-import { default as SignedLegacyTransaction } from './signedLegacyTransaction'
-import { default as SignedEIP2930Transaction } from './signedEIP2930Transaction'
+import { UnsignedLegacyTransaction, SignedLegacyTransaction } from './unsignedLegacyTransaction'
+import { UnsignedEIP2930Transaction, SignedEIP2930Transaction } from './unsignedEIP2930Transaction'
 
 /**
  * The options for initializing a Transaction.
